Move most code to codegen crate.
